[master] - Changed the encoding of instructions to fit more bytes.
NOTE - For now we only use 16 bits as MAX_REGISTERS but technically have 29 bits available.  For 29 bits we would need to allocate the register arrays in the heap and change the implementation by using a vector potentially.  This will most likely not be the case since 1 << 16 number of registers seems like an acceptable ceiling.

+++ z/object.h
@@ -13,7 +13,7 @@ typedef struct MargObject {
   const char *name;
   struct MargObject *proto;
   MargValue instance_registers[MAX_REGISTERS];
-  uint16_t instance_index;
+  uint32_t instance_index;
   EmeraldsTable instance_variables;
   EmeraldsTable messages;
   EmeraldsTable primitives;
@@ -59,7 +59,7 @@ typedef struct MargMethod {
 
   char **arguments;
   MargValue local_registers[MAX_REGISTERS];
-  uint16_t local_index;
+  uint32_t local_index;
   MargValue *constants;
   EmeraldsTable local_variables;
